pyproject.toml
[project]
name = "us-diplomatic-missions"
version = "0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "unidecode>=1.4.0",
]
